Duplicate titles use ids instead of timestamps
There was a disconnect with timestamps and checking duplicate titles.
This way, a duplicate will be distinct but a new title won't be
created each save.

Experimenting with new embed. For now Flickr is 75% working. The
functions for cleaning Facebook and Twitter are disabled due to
new way embeds are brought in. More work is needed.